Add client event callback function option to the manager
The client event callback function is called by the manager prior to an
event being sent back to a longpolling client. The callback is passed
the context for the client longpoll request and the event to be sent.
The callback can return the event as-is, modify the event, or return nil
to prevent the event from being sent.